refactor: update Icon imports to use centralized component from Utilities
